Added protocol and other pieces to allow segues between views
I followed the App Pie tutorial and got the app to move between views.
The myVCDidFinich function in the TargetWeightListController is
successfully called when you press Save on the other view and bounce
back to the first one (i added a label that changes it’s title to test
it). However, the code I wrote to append the current TargetWeightData
instance to the samplesData doesn’t seem to work, or if it is working
the view doesn’t reflect the change.