removed protocol
@benglass got the edit to work but I also re-introduced the same
problem we had before with managedObjectContext being nil and causing
an error when saving a new sample. the update one works fine so I am
not sure what’s going on. On this branch I removed the protocol that
allowed us to segue originally because I don’t think we need it with
our new coreData setup. just trying to clean things up to rule out
other problems. somehow managedObjectContext does not get initialized
in MyTableVC controller but it does in the targetList controller and
the update controller